@@ -17,19 +17,16 @@
         }
     ],
     "require": {
-        "bartacus/bartacus-bundle": "^2.0 || ^3.0",
-        "symfony/config": "^4.0 || ^5.0",
-        "symfony/dependency-injection": "^4.0 || ^5.0",
-        "symfony/http-kernel": "^4.0 || ^5.0",
-        "twig/twig": "^2.10|^3.0",
-        "typo3/cms-core": "^9.5 || ^10.4 || ^11.5",
-        "typo3/cms-extbase": "^9.5 || ^10.4 || ^11.5",
-        "typo3/cms-frontend": "^9.5 || ^10.4 || ^11.5"
-    },
-    "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.5",
-        "roave/security-advisories": "dev-master",
-        "typo3/cms-composer-installers": "^1.5 || ^2.0"
+        "php": "^8.0",
+        "composer-runtime-api": "^2.1",
+        "bartacus/bartacus-bundle": "^3.3",
+        "symfony/config": "^5.4",
+        "symfony/dependency-injection": "^5.4",
+        "symfony/http-kernel": "^5.4",
+        "twig/twig": "^3.5",
+        "typo3/cms-core": "^11.5",
+        "typo3/cms-extbase": "^11.5",
+        "typo3/cms-frontend": "^11.5"
     },
     "autoload": {
         "psr-4": {
@@ -45,6 +42,13 @@
             "extensions-in-vendor-dir": true,
             "prepare-web-dir": false,
             "extension-key": "bartacus_twig"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "typo3/cms-composer-installers": true,
+            "typo3/class-alias-loader": true,
+            "cweagans/composer-patches": true
         }
     }
 }
